rearrange map into a single component
