+++ app/views/decidim/decidim_awesome/map_component/map/show.html.erb
