+++ decidim-decidim_awesome.gemspec
@@ -18,10 +18,10 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,lib}/**/*", "LICENSE-AGPLv3.txt", "Rakefile", "README.md"]
 
-  s.add_dependency "decidim-admin", Decidim::DecidimAwesome::DECIDIM_VERSION
-  s.add_dependency "decidim-core", Decidim::DecidimAwesome::DECIDIM_VERSION
+  s.add_dependency "decidim-admin", ">= #{Decidim::DecidimAwesome::DECIDIM_VERSION}"
+  s.add_dependency "decidim-core", ">= #{Decidim::DecidimAwesome::DECIDIM_VERSION}"
   s.add_dependency "redcarpet", "~> 3.4"
 
   s.add_development_dependency "appraisal"
-  s.add_development_dependency "decidim-dev", Decidim::DecidimAwesome::DECIDIM_VERSION
+  s.add_development_dependency "decidim-dev", ">= #{Decidim::DecidimAwesome::DECIDIM_VERSION}"
 end
